Decompose the Toolbox live harness
(cherry picked from commit 6b107912dce5f66212474b501db7f91192d2ec34)
